# docs/openapi.v1.yaml
openapi: 3.1.0
info:
  title: COVA API
  version: 1.0.0
  description: |
    COVA (Collaborative Orchestrator for Verified Agents) API v1.
    This spec defines the baseline interfaces for async brief jobs and sync expert query.
servers:
  - url: https://api.cova.example.com
    description: Example production server
security:
  - bearerAuth: []
...
paths:
  /v1/assistant/jobs:
    post:
      tags: [Jobs]
      summary: Submit an async daily brief job
      operationId: submitBriefJob
      parameters:
        - $ref: '#/components/parameters/XRequestId'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/TenantIdHeader'
        - $ref: '#/components/parameters/ProjectIdHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubmitBriefRequest'
            examples:
              default:
                value:
                  expert_id: fi_cn_primary
                  expert_name: CN FI Daily Analyst
                  expert_type: fixed_income
                  date: "2026-02-27"
                  timezone: Asia/Shanghai
                  callback:
                    url: https://caller.example.com/hooks/cova
                    signing:
                      algorithm: hmac-sha256
                      secret_ref: vault://caller/cova-hook-secret
                  inference:
                    provider: openai
                    credential_mode: delegated_token
                    token: opaque-short-lived-token
                  options:
                    max_sources: 50
                    language: zh-CN
      responses:
        '202':
          description: Accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubmitBriefResponse'
        '400':
          $ref: '#/components/responses/Problem400'
        '401':
          $ref: '#/components/responses/Problem401'
        '409':
          description: Conflict caused by idempotency mismatch
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
...
  /v1/assistant/query:
    post:
      tags: [Query]
      summary: Execute a synchronous expert query
      operationId: queryExpert
      parameters:
        - $ref: '#/components/parameters/XRequestId'
        - $ref: '#/components/parameters/TenantIdHeader'
        - $ref: '#/components/parameters/ProjectIdHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QueryRequest'
            examples:
              default:
                value:
                  expert_id: fi_cn_primary
                  expert_name: CN FI Daily Analyst
                  expert_type: fixed_income
                  question: 今天利率债和信用债的主要分化是什么？
                  time_window:
                    start: "2026-02-26T00:00:00Z"
                    end: "2026-02-27T08:00:00Z"
                  inference:
                    provider: openai
                    credential_mode: delegated_token
                    token: opaque-short-lived-token
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueryResponse'
        '400':
          $ref: '#/components/responses/Problem400'
        '401':
          $ref: '#/components/responses/Problem401'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    XRequestId:
      name: X-Request-Id
      in: header
      required: false
      schema:
        type: string
      description: Caller supplied request correlation ID
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema:
        type: string
        minLength: 1
      description: Idempotency key for create operations
    TenantIdHeader:
      name: X-Tenant-Id
      in: header
      required: true
      schema:
        type: string
        minLength: 1
      description: Tenant identifier resolved from caller identity
    ProjectIdHeader:
      name: X-Project-Id
      in: header
      required: true
      schema:
        type: string
        minLength: 1
      description: Project identifier inside tenant scope
...
  responses:
    Problem400:
      description: Bad Request
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Problem401:
      description: Unauthorized
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
  schemas:
    ExpertType:
      type: string
      enum: [fixed_income, macro, equity]
    JobStatus:
      type: string
      enum: [queued, running, succeeded, failed, canceled, expired]
    SubmitBriefRequest:
      type: object
      required: [expert_id, date]
      properties:
        tenant_id:
          type: string
          description: Optional tenant echo field for auditing
        project_id:
          type: string
          description: Optional project echo field for auditing
        expert_id:
          type: string
          minLength: 1
          description: Unique expert instance ID used for routing
        expert_name:
          type: string
          description: Human readable expert name
        expert_type:
          $ref: '#/components/schemas/ExpertType'
        date:
          type: string
          pattern: '^\d{4}-\d{2}-\d{2}$'
          description: Target brief date in YYYY-MM-DD
        timezone:
          type: string
          description: IANA timezone, e.g. Asia/Shanghai
        callback:
          $ref: '#/components/schemas/CallbackConfig'
        inference:
          $ref: '#/components/schemas/InferenceConfig'
        options:
          $ref: '#/components/schemas/BriefOptions'
    CallbackConfig:
      type: object
      required: [url]
      properties:
        url:
          type: string
          format: uri
        signing:
          $ref: '#/components/schemas/CallbackSigning'
    CallbackSigning:
      type: object
      required: [algorithm, secret_ref]
      properties:
        algorithm:
          type: string
          enum: [hmac-sha256]
        secret_ref:
          type: string
          description: Secret reference in vault/secret manager
    InferenceConfig:
      type: object
      properties:
        provider:
          type: string
          example: openai
        credential_mode:
          type: string
          enum: [delegated_token]
        token:
          type: string
          description: Short-lived delegated token. Do not send long-lived API keys.
    BriefOptions:
      type: object
      properties:
        max_sources:
          type: integer
          minimum: 1
          maximum: 1000
        language:
          type: string
          example: zh-CN
    SubmitBriefResponse:
      type: object
      required: [job_id, tenant_id, project_id, expert_id, status, submitted_at]
      properties:
        job_id:
          type: string
        tenant_id:
          type: string
        project_id:
          type: string
        expert_id:
          type: string
        expert_name:
          type: string
          nullable: true
        expert_type:
          $ref: '#/components/schemas/ExpertType'
        status:
          $ref: '#/components/schemas/JobStatus'
        submitted_at:
          type: string
          format: date-time
        request_id:
          type: string
...
    QueryRequest:
      type: object
      required: [expert_id, question]
      properties:
        tenant_id:
          type: string
        project_id:
          type: string
        expert_id:
          type: string
          minLength: 1
        expert_name:
          type: string
        expert_type:
          $ref: '#/components/schemas/ExpertType'
        question:
          type: string
          minLength: 1
        time_window:
          $ref: '#/components/schemas/TimeWindow'
        inference:
          $ref: '#/components/schemas/InferenceConfig'
    QueryResponse:
      allOf:
        - $ref: '#/components/schemas/ExpertResult'
    TimeWindow:
      type: object
      required: [start, end]
      properties:
        start:
          type: string
          format: date-time
        end:
          type: string
          format: date-time
    ExpertResult:
      type: object
      required: [answer, confidence, freshness_cutoff]
      properties:
        title:
          type: string
          nullable: true
        answer:
          type: string
        confidence:
          type: number
          minimum: 0
          maximum: 1
        freshness_cutoff:
          type: string
          format: date-time
        references:
          type: array
          description: Expert-defined evidence records. Field schema is domain-specific.
          items:
            $ref: '#/components/schemas/ReferenceRecord'
        citations:
          deprecated: true
          type: array
          description: Backward-compatible fixed-income style citation list.
          items:
            $ref: '#/components/schemas/Citation'
        limitations:
          type: array
          items:
            type: string
    ReferenceRecord:
      type: object
      additionalProperties: true
    Citation:
      type: object
      required: [source_id, title, published_at]
      properties:
        source_id:
          type: string
        source_name:
          type: string
          nullable: true
        title:
          type: string
        url:
          type: string
          format: uri
          nullable: true
        published_at:
          type: string
          format: date-time
...
    Problem:
      type: object
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
          minimum: 100
          maximum: 599
        detail:
          type: string
        instance:
          type: string
          format: uri-reference
        request_id:
          type: string